django/requirements.txt
gunicorn
Django==5.0.2
django-extensions==3.2.3
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-rest-knox==4.2.0
django-filter==23.5
PyYAML==6.0.1
uritemplate==4.1.1
django-admin-interface==0.27.0
django-prometheus==2.3.1
django-redis==5.4.0
pillow>=10.2.0 # not directly required, pinned by Snyk to avoid a vulnerability
